part 1 toward animation

+++ domain/include/domain/units.h
@@ -1,6 +1,8 @@
 #pragma once
+#include <cstdint>
+
 namespace domain {
 
-using Scalar = int;
+using Scalar = int8_t;
 
 }
